fix: storybookd react native CI
