+++ apps/storybook-react-native/babel.config.js
@@ -1,8 +1,20 @@
+const resolveDesignSystemPackages = require("../../plugins/resolve-design-system-packages");
+
 module.exports = function (api) {
   api.cache(true);
 
   return {
     presets: [["babel-preset-expo", { jsxRuntime: "automatic" }]],
-    plugins: ["react-native-reanimated/plugin"],
+    plugins: [
+      "react-native-reanimated/plugin",
+      [
+        "module-resolver",
+        {
+          alias: {
+            ...resolveDesignSystemPackages(),
+          },
+        },
+      ],
+    ],
   };
 };
