+++ plugins/merged-prettier-plugin.js
